requirements.txt
accelerate==0.23.0
aiohttp==3.8.5
aiosignal==1.3.1
albumentations==1.1.0
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==2.0.12
click==8.1.6
cmake==3.27.0
coloredlogs==15.0.1
cycler==0.11.0
Cython==0.29.36
datasets==2.14.5
dill==0.3.7
docker-pycreds==0.4.0
editdistpy==0.1.3
et-xmlfile==1.1.0
evaluate==0.4.0
faiss-cpu==1.7.4
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.41.0
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.35
hdbscan==0.8.33
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
imageio==2.31.1
Jinja2==3.1.2
joblib==1.3.1
kiwisolver==1.4.4
kornia==0.6.4
linktransformer==0.1.9
lit==16.0.6
markov-clustering==0.0.6.dev0
MarkupSafe==2.1.3
matplotlib==3.5.1
memory-profiler==0.61.0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
networkx==3.1
nltk==3.8.1
numpy==1.22.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
onnx==1.13.0
onnxruntime==1.12.1
openai==0.28.0
opencv-python-headless==4.5.5.64
openpyxl==3.1.2
packaging==23.1
pandas==1.3.4
pathtools==0.1.2
Pillow==9.4.0
protobuf==3.20.3
psutil==5.8.0
pyarrow==13.0.0
pyparsing==3.1.0
pytesseract==0.3.10
python-dateutil==2.8.2
pytorch-metric-learning==1.4.0
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==6.0
qudida==0.0.4
rdflib==6.3.2
regex==2023.6.3
requests==2.31.0
responses==0.18.0
safetensors==0.3.1
scikit-image==0.19.1
scikit-learn==1.3.0
scipy==1.10.0
seaborn==0.12.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.28.1
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
sympy==1.12
symspellpy==6.7.6
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.6.2.dev0
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tqdm==4.64.1
transformers==4.31.0
triton==2.0.0
typing_extensions==4.7.1
umap==0.1.1
urllib3==1.26.16
wandb==0.15.5
xxhash==3.3.0
yarl==1.9.2
